+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- Fixed an issue with wrong patch order being generated for actions triggered inside `onAttachedToRootStore`.
 - Added support to `applyPatches` for applying patches in reverse order.
 - Fixed applying inverse patches in reverse order.